it was supposed to be better version of stats-plugin but it was too old to maintain and I need that plugin for my own server but it used to store new data only. But I wanted world stats data and other plugins compatibles too. So I recreated the plugin myself. No inspiration Just self coding ;-; took me 8hrs to fix his code but i couldn't so i made my own.
it was supposed to be better version of stats-plugin but it was too old to maintain and I need that plugin for my own server but it used to store new data only. But I wanted world stats data and other plugins compatibles too. So I recreated the plugin myself. No inspiration Just self coding ;-; took me 8hrs to fix his code but i couldn't so i made my own.